requirements.txt
streamlit==1.28.0
faiss-cpu==1.13.2
langchain==1.2.7
langchain-community==0.4.1
langchain-groq==1.1.1
langchain-text-splitters==1.1.0
sentence-transformers==2.2.2
# torch==2.10.0
huggingface-hub==0.16.4
# transformers==5.0.0
pypdf==6.6.2
python-dotenv==1.2.1
rank-bm25==0.2.2
# numpy==1.26.4
# scikit-learn==1.7.2
# scipy==1.17.0
# tqdm==4.67.1
